+++ but8.php
@@ -1,0 +1,5 @@
+<?php
+require 'functions.php';
+doThing('rpctog.php');
+if(strpos($_SERVER['HTTP_USER_AGENT'], 'WebKit')) header('Location: ' . $_SERVER['but8.php']);
+?>